Added /distributekeys >key type> <amount> to distribute keys to all players on a server.
Modified messages to let people know what they recieved.

Enabled Commands to run as a reward by adding an ItemStack Paper with name "RUN COMMAND" (case sensitive). For security reasons this will run commands in the loottable yml, so it does not grant loot table editors the ability to run commands as console.